requirements.txt
python-dotenv==1.0.0
PySimpleGUI==4.60.5
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
python-magic==0.4.27
google-generativeai==0.4.1
